mnx_canary

Module mnx

“MNX is a new, open standard for representing music notation for interchange and internal use in software applications. It builds on the ideas and success of MusicXML […]”

This code has been generated, thus modifications will be ignored.

Official documentation: https://w3c.github.io/mnx/docs/

Official schema: https://w3c.github.io/mnx/docs/mnx-schema.json

Run: cargo install cargo-typify && cargo typify mnx-schema.json

Modules

builder

Types for composing complex structures.

error

Error types.

Structs

Accent

Accent

AccidentalDisplay

AccidentalDisplay

AccidentalEnclosure

AccidentalEnclosure

Alter

Alter

Barline

Barline

Beam

Beam

BeamList

BeamList

Bpm

Bpm

BreathMark

BreathMark

BreathMarkSymbol

BreathMarkSymbol

Clef

Clef

Color

Color

Dynamic

Dynamic

DynamicList

DynamicList

DynamicType

DynamicType

Ending

Ending

EndingDuration

EndingDuration

EndingNumber

EndingNumber

EndingNumbers

EndingNumbers

EndingOpen

EndingOpen

Event

Event

EventLyricLine

EventLyricLine

EventLyricLines

EventLyricLines

EventLyricLinesKey

EventLyricLinesKey

EventMarkings

EventMarkings

Fifths

Fifths

Fine

Fine

Fraction

Fraction

Global

Global

GlobalAttrs

GlobalAttrs

Grace

Grace

GraceSequenceContent

GraceSequenceContent

Id

Id

IdList

IdList

IntegerSigned

IntegerSigned

IntegerUnsigned

IntegerUnsigned

Interval

Interval

Jump

Jump

Key

Key

Kit

Kit

KitComponent

KitComponent

KitComponentId

KitComponentId

KitKey

KitKey

KitNote

KitNote

KitNotes

KitNotes

LanguageCode

LanguageCode

LayoutChange

LayoutChange

LayoutChanges

LayoutChanges

Layouts

Layouts

LiteralStringEvent

LiteralStringEvent

LiteralStringGrace

LiteralStringGrace

LiteralStringGroup

LiteralStringGroup

LiteralStringSpace

LiteralStringSpace

LiteralStringStaff

LiteralStringStaff

LiteralStringTremolo

LiteralStringTremolo

LiteralStringTuplet

LiteralStringTuplet

LyricLineId

LyricLineId

LyricLineIdList

LyricLineIdList

LyricLineLabel

LyricLineLabel

LyricLineMetadata

LyricLineMetadata

LyricLinesMetadata

LyricLinesMetadata

LyricLinesMetadataKey

LyricLinesMetadataKey

Lyrics

Lyrics

LyricsGlobal

LyricsGlobal

MeasureCount

MeasureCount

MeasureGlobal

MeasureGlobal

MeasureNumber

MeasureNumber

MeasureRhythmicPosition

MeasureRhythmicPosition

MeasuresGlobal

MeasuresGlobal

MidiNumber

MidiNumber

Mnx

Mnx

MnxDocument

An encoding of Common Western Music Notation.

MultiNoteTremolo

MultiNoteTremolo

MultimeasureRest

MultimeasureRest

MultimeasureRests

MultimeasureRests

Note

Note

NoteValue

NoteValue

NoteValueQuantity

NoteValueQuantity

Notes

Notes

Octave

Octave

Orientation

Orientation

Ottava

Ottava

OttavaAmount

OttavaAmount

OttavaAmountOrZero

OttavaAmountOrZero

OttavaList

OttavaList

Page

Page

Pages

Pages

Part

Part

PartMeasure

PartMeasure

PartMeasures

PartMeasures

PartName

PartName

PartShortName

PartShortName

PartTransposition

PartTransposition

Parts

Parts

PerformOptions

PerformOptions

Pitch

Pitch

PositionedClef

PositionedClef

PositionedClefList

PositionedClefList

PositiveInteger

PositiveInteger

RepeatEnd

RepeatEnd

RepeatStart

RepeatStart

RepeatTimes

RepeatTimes

Rest

Rest

RhythmicPosition

RhythmicPosition

Root

Root

Score

Score

ScoreName

ScoreName

Scores

Scores

Segno

Segno

Sequence

Sequence

SequenceContent

SequenceContent

SequenceContentItem

SequenceContentItem

SequenceList

SequenceList

SimpleColor

SimpleColor

Slur

Slur

SlurList

SlurList

SlurTieEndLocation

SlurTieEndLocation

SmuflFont

SmuflFont

SmuflGlyph

SmuflGlyph

SoftAccent

SoftAccent

Sound

Sound

SoundsGlobal

SoundsGlobal

SoundsGlobalKey

SoundsGlobalKey

Space

Space

Spiccato

Spiccato

Staccatissimo

Staccatissimo

Staccato

Staccato

Staff

Staff

StaffCount

StaffCount

StaffGroup

StaffGroup

StaffLabel

StaffLabel

StaffLabelref

StaffLabelref

StaffNumber

StaffNumber

StaffPosition

StaffPosition

StaffSource

StaffSource

StaffSources

StaffSources

StressMarking

StressMarking

String

String

StrongAccent

StrongAccent

Support

Support

System

System

SystemLayout

SystemLayout

SystemLayoutContent

SystemLayoutContent

SystemLayoutContentItem

SystemLayoutContentItem

Systems

Systems

Tempo

Tempo

Tempos

Tempos

Tenuto

Tenuto

Tie

Tie

TieList

TieList

Time

Time

TimeSignatureUnit

TimeSignatureUnit

TremoloSequenceContent

TremoloSequenceContent

TremoloSingle

TremoloSingle

Tuplet

Tuplet

UnstressMarking

UnstressMarking

VendorDict

VendorDict

VendorExtensions

VendorExtensions

VendorExtensionsKey

VendorExtensionsKey

VersionNumber

VersionNumber

VoiceName

VoiceName

Written

Written

Enums

AccidentalEnclosureSymbol

AccidentalEnclosureSymbol

BarlineType

BarlineType

BeamHookDirection

BeamHookDirection

ClefSign

ClefSign

EventLyricLineType

EventLyricLineType

GraceType

GraceType

JumpType

JumpType

LineType

LineType

NoteValueBase

NoteValueBase

SlurSide

SlurSide

StaffSymbol

StaffSymbol

StemDirection

StemDirection

Step

Step

TieTargetType

TieTargetType

TimeSignatureDisplay

TimeSignatureDisplay

TupletDisplaySetting

TupletDisplaySetting

UpOrDown

UpOrDown

YesNoAuto

YesNoAuto